backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi[standard]>=0.128.5",
    "pydantic>=2.0",
    # Finance Intelligence & RAG (optional - install if using /api/finance)
    "sentence-transformers>=2.2.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "datasets>=2.14.0",
    "python-docx>=1.2.0",
    "peft>=0.18.1",
    "trl>=0.27.2",
]
[project.optional-dependencies]
finance-training = [
    "peft>=0.7.0",
    "trl>=0.7.0",
]
